+++ lib/src/otp_interactor.dart
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'dart:io';
-
 import 'package:flutter/services.dart';
 import 'package:otp_autofill/src/base/exceptions.dart';
+import 'package:otp_autofill/src/utill/platform_wrapper.dart';
 
 typedef StringCallback = void Function(String);
 
@@ -31,23 +30,32 @@ const getAppSignatureMethod = 'getAppSignature';
 
 /// Arguments.
 const senderTelephoneNumber = 'senderTelephoneNumber';
+const _defaultChannel = MethodChannel(channelName);
 
 /// Interact with native to get OTP code and telephone hint.
 class OTPInteractor {
-  static const MethodChannel _channel = MethodChannel(channelName);
+  final MethodChannel _channel;
+  final PlatformWrapper _platform;
 
   /// Show user telephone picker and get chosen number.
-  static Future<String?> get hint {
-    if (Platform.isAndroid) {
+  Future<String?> get hint {
+    if (_platform.isAndroid) {
       return _channel.invokeMethod<String>(getTelephoneHint);
     } else {
       throw UnsupportedPlatform();
     }
   }
 
+  OTPInteractor({
+    MethodChannel channel = _defaultChannel,
+    PlatformWrapper? platform,
+  })  : assert(channel.name == channelName),
+        _channel = channel,
+        _platform = platform ?? PlatformWrapper();
+
   /// Get app signature, that used in Retriever API.
-  static Future<String?> getAppSignature() async {
-    if (Platform.isAndroid) {
+  Future<String?> getAppSignature() async {
+    if (_platform.isAndroid) {
       return _channel.invokeMethod<String>(getAppSignatureMethod);
     } else {
       throw UnsupportedPlatform();
@@ -61,7 +69,7 @@ class OTPInteractor {
 
   /// Broadcast receiver start listen for OTP code with User Consent API.
   Future<String?> startListenUserConsent([String? senderPhone]) async {
-    if (Platform.isAndroid) {
+    if (_platform.isAndroid) {
       return _channel.invokeMethod<String>(
         startListenUserConsentMethod,
         <String, String?>{
@@ -75,7 +83,7 @@ class OTPInteractor {
 
   /// Broadcast receiver start listen for OTP code with Retriever API.
   Future<String?> startListenRetriever() async {
-    if (Platform.isAndroid) {
+    if (_platform.isAndroid) {
       return _channel.invokeMethod<String>(startListenRetrieverMethod);
     } else {
       throw UnsupportedPlatform();