+++ lib/src/utill/platform_wrapper.dart
@@ -1,0 +1,6 @@
+import 'dart:io';
+
+/// Wrapper for Platform io.
+class PlatformWrapper {
+  bool get isAndroid => Platform.isAndroid;
+}